// src/models/interface.ts
export type Id = mongoose.Types.ObjectId;
// ...
export type Group =
  | "A"
  | "B"
  | "C"
  | "Dog"
  | "E"
  | "F"
  | "G"
  | "H"
  | "J"
  | "K"
  | "L"
  | "M"
  | "N"
  | "P"
  | "Q"
  | "R"
  | "S"
  | "T";
export type Size = "S" | "M" | "L" | "XL" | "XXL" | "3XL";
export type RoleCamp = Mode | "peto";
export type Role = RoleCamp | "admin";
export type Mode = "nong" | "pee";
export interface HeathIssueBody {
  food: string;
  chronicDisease: string;
  medicine: string;
  extra: string;
  isWearing: boolean;
  spicy: boolean;
  foodConcern: string;
  foodLimit: FoodLimit;
  //private
}
// ...
export interface UpdateTimeOffsetRaw {
  day: number;
  hour: number;
  minute: number;
  //private
}
// ...
export const foodLimits = [
  "อิสลาม",
  "มังสวิรัติ",
  "เจ",
  "ไม่มีข้อจำกัดด้านความเชื่อ",
] as const;
export type FoodLimit = (typeof foodLimits)[number];
export interface HeathIssuePack {
  heathIssue: HeathIssueBody;
  user: BasicUser;
  campMemberCardId: Id;
  //private
}
// ...
export interface InterFood {
  campId: Id;
  isWhiteList: boolean;
  nongCampMemberCardIds: Id[];
  peeCampMemberCardIds: Id[];
  petoCampMemberCardIds: Id[];
  name: string;
  mealId: Id;
  lists: FoodLimit[];
  _id: Id;
  isSpicy: boolean;
  listPriority: boolean;
  //public
}
export interface InterMeal {
  time: Date;
  campId: Id;
  foodIds: Id[];
  roles: RoleCamp[];
  _id: Id;
  //public
}
export interface CreateFood {
  campId: Id;
  isWhiteList: boolean;
  name: string;
  mealId: Id;
  lists: FoodLimit[];
  isSpicy: boolean;
  listPriority: boolean;
  //public
}
export interface CreateMeal {
  time: Date;
  campId: Id;
  roles: RoleCamp[];
  //public
}
export interface UpdateFood {
  isWhiteList: boolean;
  nongCampMemberCardIds: Id[];
  peeCampMemberCardIds: Id[];
  petoCampMemberCardIds: Id[];
  name: string;
  lists: FoodLimit[];
  _id: Id;
  isSpicy: boolean;
  listPriority: boolean;
  //public
}
export interface GetFoodForUpdate {
  isWhiteList: boolean;
  nongHealths: HeathIssuePack[];
  peeHealths: HeathIssuePack[];
  petoHealths: HeathIssuePack[];
  nongCampMemberCardIds: Id[];
  peeCampMemberCardIds: Id[];
  petoCampMemberCardIds: Id[];
  name: string;
  lists: FoodLimit[];
  _id: Id;
  isSpicy: boolean;
  camp: BasicCamp;
  time: Date;
  listPriority: boolean;
  displayOffset: UpdateTimeOffsetRaw;
  //public
}
export interface GetMeals {
  time: Date;
  whiteLists: InterFood[];
  blackLists: InterFood[];
  //private
}
// ...
export interface UpdateMeal {
  mealId: Id;
  time: Date;
  roles: RoleCamp[];
  //public
}
// ...
export interface BasicUser {
  _id: Id;
  name: string;
  lastname: string;
  nickname: string;
  gender: "Male" | "Female";
  shirtSize: Size;
  haveBottle: boolean;
  mode: Mode;
  group: Group | null;
  role: Role;
  fridayActEn: boolean;
  fridayAuth: boolean;
  likeToSleepAtCamp: boolean;
  //private
}
export interface BasicBaan {
  name: string;
  fullName: string | null;
  campId: Id;
  link: string | null;
  _id: Id;
  mdTime: Date;
  nongSendMessage: boolean;
  songIds: Id[];
  canReadMirror: boolean;
  canWriteMirror: boolean;
  mirrorIds: Id[];
  nongIds: Id[];
  peeIds: Id[];
  //public
}
export interface BasicCamp {
  nameId: Id;
  round: number;
  dateStart: Date;
  dateEnd: Date;
  nongDataLock: boolean;
  open: boolean;
  songIds: Id[];
  link: string | null;
  allDone: boolean;
  lockChangePickup: boolean;
  pictureUrls: string[];
  registerModel: "noPaid" | "noInterview" | "all";
  memberStructure:
    | "nong->highSchool,pee->1year,peto->2upYear"
    | "nong->highSchool,pee->2upYear"
    | "nong->1year,pee->2upYear"
    | "nong->highSchool,pee->allYear"
    | "allYearMix";
  logoUrl: string | null;
  registerSheetLink: string | null;
  peeLock: boolean;
  _id: Id;
  campName: string;
  nongSleepModel:
    | "นอนทุกคน"
    | "เลือกได้ว่าจะค้างคืนหรือไม่"
    | "ไม่มีการค้างคืน";
  peeSleepModel: "นอนทุกคน" | "เลือกได้ว่าจะค้างคืนหรือไม่" | "ไม่มีการค้างคืน";
  groupName: string;
  peeDataLock: boolean;
  petoDataLock: boolean;
  haveCloth: boolean;
  actionPlanOffset: number;
  mdTime: Date;
  showCorrectAnswerAndScore: boolean;
  canAnswerTheQuestion: boolean;
  mealIds: Id[];
  canNongSeeAllAnswer: boolean;
  canNongSeeAllActionPlan: boolean;
  canNongSeeAllTrackingSheet: boolean;
  canNongAccessDataWithRoleNong: boolean;
  lockChangeQuestion: boolean;
  canReadTimeOnMirror: boolean;
  nongCall: string;
  boyZoneLadyZoneState: BoyZoneLadyZoneState;
  canNongSeeBaanOrder: boolean;
  nongIds: Id[];
  peeIds: Id[];
  petoIds: Id[];
  boardIds: Id[];
  baanIds: Id[];
  partIds: Id[];
  //public
}
// ...
export type BoyZoneLadyZoneState = (typeof boyZoneLadyZoneStates)[number];